bb198

+++ src/features/create-post/steps/add-photo-slider/add-photo-slider.tsx
@@ -3,10 +3,12 @@ import React, { ChangeEvent, useRef } from 'react'
 import { Popover } from '@headlessui/react'
 import Image from 'next/image'
 import { toast, Toaster } from 'react-hot-toast'
+import { useDispatch } from 'react-redux'
 
 import style from './add-photo-slider.module.scss'
 
 import { CropContextType } from '@/features/create-post/context/crop-provider'
+import { setPhotosCount } from '@/shared/api/services/posts/post.slice'
 import addPhoto from '@/shared/assets/icons/add-photo/add-photo.svg'
 import noImage from '@/shared/assets/icons/image/no-image.svg'
 import { Button, ButtonTheme } from '@/shared/ui'
@@ -17,8 +19,9 @@ type Props = {
 }
 
 export const AddPhotoSlider = ({ cropContext, setCurrentIndex }: Props) => {
+  const dispatch = useDispatch()
   const inputRef = useRef<HTMLInputElement>(null)
-
+  const photosLength = cropContext.photos.length
   const handlerAddImageClick = async (event: ChangeEvent<HTMLInputElement>) => {
     const files = event.target.files
 
@@ -53,7 +56,12 @@ export const AddPhotoSlider = ({ cropContext, setCurrentIndex }: Props) => {
     inputRef.current?.click()
   }
   const handleDeleteClick = (index: number) => {
-    cropContext.deletePhoto(index)
+    if (photosLength > 1) {
+      cropContext.deletePhoto(index)
+      dispatch(setPhotosCount(photosLength - 1))
+    } else {
+      cropContext.setIsOpenModal(true)
+    }
   }
   const isButtonDisabled = cropContext.photos.length >= 10
